Dependences.py
#Math and tables
import numpy as np
import pandas as pd
from scipy.optimize import linear_sum_assignment

# Ultralytics
from ultralytics import YOLO
from ultralytics.nn.tasks import DetectionModel
from ultralytics.utils.metrics import bbox_iou
from ultralytics.utils.loss import v8DetectionLoss
from ultralytics.utils.ops import non_max_suppression
from ultralytics.engine.results import Results, Boxes
from ultralytics.utils.metrics import OKS_SIGMA
from ultralytics.utils.ops import crop_mask, xywh2xyxy, xyxy2xywh
from ultralytics.utils.tal import RotatedTaskAlignedAssigner, TaskAlignedAssigner, dist2bbox, dist2rbox, make_anchors
from ultralytics.utils.loss import BboxLoss
from ultralytics.utils.metrics import compute_ap

#Pipilines
from transformers import pipeline

#Losses and metrics
from pytorch_msssim import ssim 
from skimage.metrics import peak_signal_noise_ratio

#Pytorch
import torch
from torch import nn
from torch.utils.data import DataLoader, Dataset, random_split 
import torch.utils.checkpoint as checkpoint
from torch.linalg import svd
import torch.optim as optim


#Transforms
import torchvision
from torchvision import transforms
from torchvision import transforms as tr

#Image preprocessing and visualization
from transformers import AutoImageProcessor, AutoModelForDepthEstimation, pipeline
from PIL import Image
import matplotlib.pyplot as plt
import cv2

#Logging
import wandb

#Basic python libs
import yaml
import gc
import pickle
import random
import warnings
import json
import os

