requirements.txt
matplotlib==3.10.3
numpy==2.3.0
pandas==2.3.0
Pillow==11.2.1
pytorch_msssim==1.0.0
PyYAML==6.0.2
PyYAML==6.0.2
scipy==1.15.3
skimage==0.0
torch==2.7.1+cpu
torchvision==0.22.1+cpu
transformers==4.52.4
ultralytics==8.3.156
wandb==0.20.1
